requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
datasets==4.0.0
dill==0.3.8
Django==5.2.6
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.30.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
language_tool_python==2.9.4
Markdown==3.9
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
nltk==3.9.1
numpy==2.3.2
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyphen==0.17.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentencepiece==0.2.1
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
textblob==0.19.0
textstat==0.7.10
threadpoolctl==3.6.0
tokenizers==0.20.3
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.45.2
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.9.0
xxhash==3.5.0
yarl==1.20.1
